docs(azure): merge azure and ubuntu security features doc
Discussions with respect to [0][1] determined that the full content of
the Azure Security Overview document was unsuitable, as the majority of
its content was non-specific to Azure and only reiterated security
features and products core to Ubuntu - and not specific to Azure.
Instead, merge the relevant sections (Ubuntu Security Features and
Azure Security Features) into the existing security type document.

References:

[0] https://github.com/canonical/ubuntu-cloud-docs/pull/269
[1] https://warthogs.atlassian.net/browse/CPC-4950