feat(gce): Update Pro upgrade instructions
Google is rolling out mutable licenses which allow changing the license
of the boot disk. Google has asked us to use this new mechanism for
in-place upgrades to Pro.

What current instructions describe is appending Pro license to the
`userLicenses` field of the boot disk, but not changing its `license`
field. The new approach modifies the `license` field.

In addition to describe new mechanisms, some style changes were made:

  1. consistent variable references (FOO vs $FOO, and monospace)

  2. formatting of lists

  3. make the content below numbered list items actually belong to those
     items

[1] https://community.cochrane.org/style-manual/formatting/lists