feat(oracle): Update OKE to reflect image usage for LA2
We are approaching LA2 for OKE and how customers can list and use images has changed.
Instead of downloading and registering the image themselves, they now can list them
from the Oracle API and use them directly.

The docs have been changed to reflect that and discourge the previous process
of manually downloading and registering themselves.