feat(azure): Jammy unified offer
In an effort to align with the offer structure we have employed from
24.04 forward, we are introducing a unified 22.04 offer. Update the
documentation to point to this offer instead. The old offers still
exist, will receive updates and can be found via the CLI.

This commit does not update the Jammy non-Pro minimal links since
publication issues have prevented that SKU from being published.