fix(azure): Consistent table sizing
Column widths are relative to each other, but to stay consistent, all
column widths in a table now add to 100.

Note, if a single word is longer than the relative length of that
column, the column will be larger than the specified relative width.
Becuase of this, changed column names to be shorter. This limits blank
space between columns and gives more length to the URNs, which commonly
wrap to the next line.

Words like Hyper-V are treated as 2 words, so the relative width of that
column was chosen to avoid wrapping the '-V' onto the next line.