feat(azure): 22.04 and newer golden images
Updates the guide for creating golden images to work for Ubuntu 22.04
and newer. There are differences in the underlying offers using from
22.04 onward which require some changes to the instructions.

The newer offers do not contain publisher/offer/sku information in the
attested metadata and therefore auto-attaching to Pro is not
possible. For these images we update the instructions to not detach
from Pro during the build to ensure the final image has Pro
entitlements. The user experience for these images is otherwise
unchanged.